+++ examples/make_points_example_config.py
@@ -5,7 +5,7 @@ import numpy as np
 from cellier.models.data_manager import DataManager
 from cellier.models.data_stores.points import PointsMemoryStore
 from cellier.models.data_streams.points import PointsSynchronousDataStream
-from cellier.models.nodes.points_visual import PointsNode, PointsUniformMaterial
+from cellier.models.nodes.points_node import PointsNode, PointsUniformMaterial
 from cellier.models.scene.cameras import PerspectiveCamera
 from cellier.models.scene.canvas import Canvas
 from cellier.models.scene.dims_manager import CoordinateSystem, DimsManager
