+++ src/cellier/models/nodes/mesh_node.py
