+++ src/cellier/models/nodes/points_node.py
